+++ ImageProcessing/vtkCudaImageAnalytics/vtkCudaImageAtlasLabelProbability.cxx
@@ -86,17 +86,17 @@ void vtkCudaImageAtlasLabelProbability::SetOutputToProbability()
 //----------------------------------------------------------------------------
 void vtkCudaImageAtlasLabelProbability::SetStDevX(double val)
 {
-  this->GaussianDevs[0] = (val > 0.0f ? val: 0.0f);
+  this->GaussianDevs[0] = (val > 0.0f ? val : 0.0f);
 }
 
 void vtkCudaImageAtlasLabelProbability::SetStDevY(double val)
 {
-  this->GaussianDevs[1] = (val > 0.0f ? val: 0.0f);
+  this->GaussianDevs[1] = (val > 0.0f ? val : 0.0f);
 }
 
 void vtkCudaImageAtlasLabelProbability::SetStDevZ(double val)
 {
-  this->GaussianDevs[2] = (val > 0.0f ? val: 0.0f);
+  this->GaussianDevs[2] = (val > 0.0f ? val : 0.0f);
 }
 
 double vtkCudaImageAtlasLabelProbability::GetStDevX()
@@ -116,44 +116,44 @@ double vtkCudaImageAtlasLabelProbability::GetStDevZ()
 
 //----------------------------------------------------------------------------
 // The output extent is the intersection.
-int vtkCudaImageAtlasLabelProbability::RequestInformation (
-  vtkInformation * vtkNotUsed(request),
-  vtkInformationVector **inputVector,
-  vtkInformationVector *outputVector)
+int vtkCudaImageAtlasLabelProbability::RequestInformation(
+  vtkInformation* vtkNotUsed(request),
+  vtkInformationVector** inputVector,
+  vtkInformationVector* outputVector)
 {
   // get the info objects
   vtkInformation* outInfo = outputVector->GetInformationObject(0);
-  vtkInformation *inInfo = inputVector[0]->GetInformationObject(0);
+  vtkInformation* inInfo = inputVector[0]->GetInformationObject(0);
   vtkDataObject::SetPointDataActiveScalarInfo(outInfo, VTK_FLOAT, 1);
 
   int numLabelMaps = 0;
-  for(int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
+  for (int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
   {
     numLabelMaps++;
   }
 
   int ext[6], ext2[6], idx;
 
-  inInfo->Get(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(),ext);
+  inInfo->Get(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(), ext);
 
-  for(int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
+  for (int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
   {
-    vtkInformation *inInfo2 = inputVector[0]->GetInformationObject(i);
-    inInfo->Get(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(),ext2);
+    vtkInformation* inInfo2 = inputVector[0]->GetInformationObject(i);
+    inInfo->Get(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(), ext2);
     for (idx = 0; idx < 3; ++idx)
     {
-      if (ext2[idx*2] > ext[idx*2])
+      if (ext2[idx * 2] > ext[idx * 2])
       {
-        ext[idx*2] = ext2[idx*2];
+        ext[idx * 2] = ext2[idx * 2];
       }
-      if (ext2[idx*2+1] < ext[idx*2+1])
+      if (ext2[idx * 2 + 1] < ext[idx * 2 + 1])
       {
-        ext[idx*2+1] = ext2[idx*2+1];
+        ext[idx * 2 + 1] = ext2[idx * 2 + 1];
       }
     }
   }
 
-  outInfo->Set(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(),ext,6);
+  outInfo->Set(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(), ext, 6);
 
   return 1;
 }
@@ -164,18 +164,18 @@ int vtkCudaImageAtlasLabelProbability::RequestInformation (
 // It just executes a switch statement to call the correct function for
 // the datas data types.
 int vtkCudaImageAtlasLabelProbability::RequestData(
-  vtkInformation * vtkNotUsed( request ),
-  vtkInformationVector ** inputVector,
-  vtkInformationVector * outputVector )
+  vtkInformation* vtkNotUsed(request),
+  vtkInformationVector** inputVector,
+  vtkInformationVector* outputVector)
 {
 
   vtkImageData** inData = new vtkImageData*[inputVector[0]->GetNumberOfInformationObjects()];
 
   int numLabelMaps = 0;
-  for(int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
+  for (int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
   {
     inData[i] = vtkImageData::SafeDownCast(inputVector[0]->GetInformationObject(i)->Get(vtkDataObject::DATA_OBJECT()));
-    if(inData[i])
+    if (inData[i])
     {
       numLabelMaps++;
     }
@@ -183,7 +183,7 @@ int vtkCudaImageAtlasLabelProbability::RequestData(
 
   if (numLabelMaps == 0)
   {
-    vtkErrorMacro( "At least one label map is required." );
+    vtkErrorMacro("At least one label map is required.");
     return -1;
   }
 
@@ -194,30 +194,30 @@ int vtkCudaImageAtlasLabelProbability::RequestData(
   // this filter expects the output datatype to be float.
   if (outData->GetScalarType() != VTK_FLOAT)
   {
-    vtkErrorMacro( "Output data type must be float." );
+    vtkErrorMacro("Output data type must be float.");
     return -1;
   }
 
   // this filter expects the label map to be of type char
   int LabelType = -1;
-  for(int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
+  for (int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
   {
-    if( !inData[i] )
+    if (!inData[i])
     {
       continue;
     }
-    if( LabelType == -1 )
+    if (LabelType == -1)
     {
       LabelType = inData[i]->GetScalarType();
     }
     if (inData[i]->GetScalarType() != LabelType)
     {
-      vtkErrorMacro( "Label maps must be of same type." );
+      vtkErrorMacro("Label maps must be of same type.");
       return -1;
     }
-    if ( inData[i]->GetNumberOfScalarComponents() != 1 )
+    if (inData[i]->GetNumberOfScalarComponents() != 1)
     {
-      vtkErrorMacro( "Label map can only have 1 component." );
+      vtkErrorMacro("Label map can only have 1 component.");
       return -1;
 
     }
@@ -229,13 +229,13 @@ int vtkCudaImageAtlasLabelProbability::RequestData(
   //allocate some working buffers
   short* agreementGPU;
   float* outputGPU;
-  CUDA_GetRelevantBuffers(&agreementGPU,&outputGPU,VolumeSize,this->GetStream());
+  CUDA_GetRelevantBuffers(&agreementGPU, &outputGPU, VolumeSize, this->GetStream());
 
   //figure out the agreement
   short Number = 0;
-  for(int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
+  for (int i = 0; i < inputVector[0]->GetNumberOfInformationObjects(); i++)
   {
-    if( !inData[i] )
+    if (!inData[i])
     {
       continue;
     }
@@ -243,11 +243,11 @@ int vtkCudaImageAtlasLabelProbability::RequestData(
     this->ReserveGPU();
     switch (inData[i]->GetScalarType())
     {
-      vtkTemplateMacro(
-        CUDA_IncrementInformation((VTK_TT *) inData[i]->GetScalarPointer(), (VTK_TT) LabelID, agreementGPU, VolumeSize,this->GetStream()));
-    default:
-      vtkErrorMacro( "Execute: Unknown ScalarType");
-      return -1;
+        vtkTemplateMacro(
+          CUDA_IncrementInformation((VTK_TT*) inData[i]->GetScalarPointer(), (VTK_TT) LabelID, agreementGPU, VolumeSize, this->GetStream()));
+      default:
+        vtkErrorMacro("Execute: Unknown ScalarType");
+        return -1;
     }
   }
 
@@ -255,10 +255,15 @@ int vtkCudaImageAtlasLabelProbability::RequestData(
   short flags = this->Entropy ? 1 : 0;
   flags += this->NormalizeDataTerm ? 2 : 0;
   flags += this->GaussianBlurOn ? 4 : 0;
-  int GaussWidth[3] = {this->GaussianDevs[0]*this->GaussianDevs[0] * 4.0, this->GaussianDevs[1]*this->GaussianDevs[1] * 4.0, this->GaussianDevs[2]*this->GaussianDevs[2] * 4.0 };
+  int GaussWidth[3] =
+  {
+    static_cast<int>(this->GaussianDevs[0]* this->GaussianDevs[0] * 4.0),
+    static_cast<int>(this->GaussianDevs[1]* this->GaussianDevs[1] * 4.0),
+    static_cast<int>(this->GaussianDevs[2]* this->GaussianDevs[2] * 4.0)
+  };
   this->ReserveGPU();
-  CUDA_ConvertInformation(agreementGPU, outputGPU, this->MaxValueToGive, VolumeSize, Number, flags, GaussWidth, outData->GetDimensions(), this->GetStream() );
-  CUDA_CopyBackResult(outputGPU, (float*) outData->GetScalarPointer(), VolumeSize, this->GetStream() );
+  CUDA_ConvertInformation(agreementGPU, outputGPU, this->MaxValueToGive, VolumeSize, Number, flags, GaussWidth, outData->GetDimensions(), this->GetStream());
+  CUDA_CopyBackResult(outputGPU, (float*) outData->GetScalarPointer(), VolumeSize, this->GetStream());
 
   //deallocate memory
   delete[] inData;
@@ -270,7 +275,7 @@ int vtkCudaImageAtlasLabelProbability::RequestData(
 int vtkCudaImageAtlasLabelProbability::FillInputPortInformation(
   int port, vtkInformation* info)
 {
-  info->Set(vtkAlgorithm::INPUT_IS_REPEATABLE(),1);
+  info->Set(vtkAlgorithm::INPUT_IS_REPEATABLE(), 1);
   info->Set(vtkAlgorithm::INPUT_REQUIRED_DATA_TYPE(), "vtkImageData");
   return 1;
 }
@@ -278,15 +283,15 @@ int vtkCudaImageAtlasLabelProbability::FillInputPortInformation(
 //----------------------------------------------------------------------------
 void vtkCudaImageAtlasLabelProbability::PrintSelf(ostream& os, vtkIndent indent)
 {
-  this->Superclass::PrintSelf(os,indent);
+  this->Superclass::PrintSelf(os, indent);
 
 }
 
 //----------------------------------------------------------------------------
-void vtkCudaImageAtlasLabelProbability::SetInputLabelMapConnection(vtkAlgorithmOutput *in, int number)
+void vtkCudaImageAtlasLabelProbability::SetInputLabelMapConnection(vtkAlgorithmOutput* in, int number)
 {
-  if(number >= 0)
+  if (number >= 0)
   {
-    this->SetNthInputConnection(0,number,in);
+    this->SetNthInputConnection(0, number, in);
   }
 }
